separate out scope map to it's own module
